코드 구조 변경 (가독성 증가) 및 exception 처리 코드 변경

+++ src/main/java/com/playdata/comment/service/CommentService.java
@@ -7,7 +7,9 @@ import com.playdata.domain.article.repository.ArticleRepository;
 import com.playdata.domain.comment.entity.Comment;
 import com.playdata.domain.comment.repository.CommentRepository;
 import com.playdata.domain.comment.request.CommentRequest;
+import com.playdata.domain.comment.response.CommentResponse;
 import com.playdata.exception.NoArticleByIdException;
+import com.playdata.exception.NoCommentByIdException;
 import com.playdata.exception.NotCorrectTokenIdException;
 import lombok.RequiredArgsConstructor;
 
@@ -29,31 +31,35 @@ public class CommentService {
     public void insertComment(CommentRequest commentRequest, Long articleId, UUID memberId)
     {
         Optional<Article> getArticleById = articleRepository.findByIdAndAndDeletedAtIsNull(articleId);
-        getArticleById.orElseThrow(()->new NoArticleByIdException("No Article"));
+        getArticleById.orElseThrow(()->
+                new NoArticleByIdException("No Article . id = {%s}".formatted(String.valueOf(articleId))));
         commentRepository.save(commentRequest.toEntity(memberId, articleId));
     }
-    @Transactional
-    public void deleteComment(TokenInfo tokenInfo, Long id)
-    {
-        Optional<Comment> getCommentById = commentRepository.findById(id);
-        Comment comment=getCommentById.orElseThrow(()-> new NoArticleByIdException("No Comment"));
-        if(tokenInfo.getId().equals(comment.getMember().getId())) {
-            comment.delete();
-        }
-        else {
-            new NotCorrectTokenIdException("Not Correct Token");
-        }
-    }
-    @Transactional
-    public Comment updateComment(TokenInfo tokenInfo,Long id, CommentRequest commentRequest)
-    {
-        Optional<Comment> getCommentById = commentRepository.findById(id);
-        Comment comment=getCommentById.orElseThrow(()->new NoArticleByIdException("No Comment"));
-        if(tokenInfo.getId().equals(comment.getMember().getId())) comment.setContent(commentRequest.content());
-        else {
-            new NotCorrectTokenIdException(("Not Correct Token"));
-        }
+    public Comment findCommentById(Long id){
+        Optional<Comment> getCommentById =commentRepository.findById(id);
+        Comment comment= getCommentById.orElseThrow(()->
+                new NoCommentByIdException("No Comment . id = {%s}".formatted(String.valueOf(id))));
         return comment;
+    }
+    
+    @Transactional
+    public void deleteComment(TokenInfo tokenInfo, Long id) throws NotCorrectTokenIdException {
+        Comment comment=findCommentById(id);
+        if(checkTokenAvailabiltiy(tokenInfo, comment)) comment.delete();
+    }
+
+    public boolean checkTokenAvailabiltiy(TokenInfo tokenInfo, Comment comment) throws NotCorrectTokenIdException {
+        if(!tokenInfo.getId().equals(comment.getMember().getId()))
+            throw new NotCorrectTokenIdException("Not Correct Token . token id = {%s}"
+                    .formatted(String.valueOf(tokenInfo.getId())));
+        return true;
+    }
+
+    @Transactional
+    public CommentResponse updateComment(TokenInfo tokenInfo, Long id, CommentRequest commentRequest) throws NotCorrectTokenIdException {
+        Comment comment=findCommentById(id);
+        if(checkTokenAvailabiltiy(tokenInfo, comment)) comment.setContent(commentRequest.content());
+        return new CommentResponse(comment);
     }
